test: Listener, 통합 테스트 병합 및 이벤트 발행 방식으로 변경
